same

+++ sys-kernel/genpnprd/files/rmmod.c
@@ -1,4 +1,4 @@
-extern int delete_module(const char *name, int flags);
+extern long delete_module(const char *name, int flags);
 int main(int argc, char *argv[]){
 	int i;
 	for (i=1; i<argc; i++) delete_module(argv[i], 0);
